add jaxtyping import hook. rename src directory to ppo_brax_equinox. add __init__.py
